step 4A

+++ src/main/java/com/miraleva/ceng301/repository/EnrollmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnrollmentRepository extends JpaRepository<EnrollmentEntity, Long> {
+public interface EnrollmentRepository extends JpaRepository<EnrollmentEntity, Integer> {
 }
